+++ Src/Core/VpnHood.Core.Client/ConnectorServices/ConnectorServiceBase.cs
@@ -283,9 +283,6 @@ internal class ConnectorServiceBase : IDisposable
     //private bool UserCertificateValidationCallback2(object sender, X509Certificate? certificate, X509Chain? chain,
     //    SslPolicyErrors sslPolicyErrors)
     //{
-    //    if (certificate == null!) // for android 6 (API 23)
-    //        return true;
-
     //    var ret = sslPolicyErrors == SslPolicyErrors.None ||
     //              EndPointInfo.CertificateHash?.SequenceEqual(certificate.GetCertHash()) == true;